Move the rpy2-specific code to separate files
